Struct currency::Currency

pub struct Currency {
    // some fields omitted
}

Represents currency through an optional symbol and amount of coin.

Every 100 coins represents a banknote. (coin: 100 => 1.00)

Methods

impl Currency

fn new() -> Self

Creates a blank Currency with no symbol and 0 coin.

fn from_str(s: &str) -> Result<Currency, ParseCurrencyError>

Parses a string literal (&str) and attempts to convert it into a currency. Returns Ok(Currency) on a successful conversion, otherwise Err(ParseCurrencyError).

Examples

use currency::Currency;

let c1 = Currency::from_str("$42.32").unwrap();
let c2 = Currency::from_str("$0.10").unwrap();
assert_eq!(c1 + c2, Currency::from_str("$42.42").unwrap());

fn sign(&self) -> Sign

Returns the Sign of the BigInt holding the coins.

fn value(&self) -> &BigInt

Returns the number of coins held in the Currency as &BigInt.

Should you need ownership of the returned BigInt, call clone() on it.

Examples

extern crate num;
extern crate currency;

fn main() {
    use num::traits::ToPrimitive;
    use currency::Currency;

    let c1 = Currency::new();
    assert_eq!(c1.value().to_u32().unwrap(), 0);

    let c2 = Currency::from_str("$1.42").unwrap();
    assert_eq!(c2.value().to_u32().unwrap(), 142);
}

fn convert(&self, conversion_rate: f64, currency_symbol: char) -> Currency

Returns a new Currency by multiplying the coin by the conversion rate and changing the symbol.

Examples

use currency::Currency;

let dollars = Currency::from_str("$10.00").unwrap();
let conv_rate = 0.89;
let euros = dollars.convert(0.89, '€');
assert_eq!(euros, Currency::from_str("€8.90").unwrap());