Struct currency::Currency

pub struct Currency(pub Option<char>, pub i64);

Represents currency through an optional symbol and amount of coin.

Each 100 coins results in a banknote. (100 is formatted as 1.00) The currency will be formatted as such: Currency(Some('$'), 432) ==> "$4.32"

Methods

impl Currency

fn new() -> Currency

Creates a blank Currency as Currency(None, 0)

Examples

let mut c = Currency::new();

fn from_string(s: &str) -> Currency

Parses a string literal and turns it into a currency.

Parsing ignores spaces and commas, only taking note of the digits and leading sign.

Examples

Currency::from_string("$4.32") -> Currency(Some('$'), 432)
Currency::from_string("424.44") -> Currency(None, 42444)
Currency::from_string("@12") -> Currency(Some('@'), 1200)

Failures

Fails to take note of the floating points position. Currency::from_string("$42.012) -> Currency(Some('$'), 42012) Currency::from_string("42.") -> Currency(None, 42)

Panics

Panics if a number fails to be parsed; this only occurs if the string argument has no numbers in it.

Safety

If a decimal point is intended to be marked, always use '.' A "European style" ',' will be ignored. String::from_string("€4.32") instead of String::from_string("€4,32")